# docs/openapi/zerg.v1.json
{
  "components": {
...
    "schemas": {
...
      "CreateTask": {
        "description": "DTO for creating a new task",
        "properties": {
          "description": {
            "type": "string"
          },
          "due_date": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "priority": {
            "$ref": "#/components/schemas/TaskPriority"
          },
          "project_id": {
            "format": "uuid",
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "type": "object"
      },
...
      "Task": {
        "description": "Task entity - represents a task",
        "properties": {
          "completed": {
            "description": "Whether the task is completed",
            "type": "boolean"
          },
          "created_at": {
            "description": "Creation timestamp",
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "description": "Task description",
            "type": "string"
          },
          "due_date": {
            "description": "Optional due date",
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "description": "Unique identifier",
            "format": "uuid",
            "type": "string"
          },
          "org_ref": {
            "description": "Owning tenant, as an identity-provider reference (`org_01…` or\n`personal:{subject}`). Not a foreign key - the tasks service owns its own\ndatabase and never joins against zerg's `organizations` table.",
            "type": "string"
          },
          "priority": {
            "$ref": "#/components/schemas/TaskPriority",
            "description": "Task priority"
          },
          "project_id": {
            "description": "Optional project association",
            "format": "uuid",
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus",
            "description": "Task status"
          },
          "title": {
            "description": "Task title",
            "type": "string"
          },
          "updated_at": {
            "description": "Last update timestamp",
            "format": "date-time",
            "type": "string"
          },
          "user_ref": {
            "description": "Owning user, as an identity-provider reference (`user_01…`).",
            "type": "string"
          }
        },
        "required": [
          "id",
          "org_ref",
          "user_ref",
          "title",
          "description",
          "completed",
          "priority",
          "status",
          "created_at",
          "updated_at"
        ],
        "type": "object"
      },
      "TaskPriority": {
        "description": "Task priority levels",
        "enum": [
          "low",
          "medium",
          "high",
          "urgent"
        ],
        "type": "string"
      },
      "TaskStatus": {
        "description": "Task status",
        "enum": [
          "todo",
          "in_progress",
          "done"
        ],
        "type": "string"
      },
...
      "UpdateTask": {
        "description": "DTO for updating an existing task",
        "properties": {
          "completed": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "due_date": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "priority": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/TaskPriority"
              },
              {
                "type": "null"
              }
            ]
          },
          "project_id": {
            "format": "uuid",
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/TaskStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "session_cookie": {
        "in": "cookie",
        "name": "zerg_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "API for managing tasks, projects, cloud resources, and users",
    "license": {
      "name": ""
    },
    "title": "Zerg API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/tasks": {
      "get": {
        "operationId": "list_tasks",
        "parameters": [
          {
            "in": "query",
            "name": "mine",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "project_id",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/TaskStatus"
                },
                {
                  "type": "null"
                }
              ]
            }
          },
          {
            "in": "query",
            "name": "priority",
            "required": false,
            "schema": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/TaskPriority"
                },
                {
                  "type": "null"
                }
              ]
            }
          },
          {
            "in": "query",
            "name": "completed",
            "required": false,
            "schema": {
              "type": [
                "boolean",
                "null"
              ]
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Task"
                  },
                  "type": "array"
                }
              }
            },
            "description": "List of tasks"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ],
        "summary": "List tasks",
        "tags": [
          "tasks"
        ]
      },
      "post": {
        "operationId": "create_task",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTask"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            },
            "description": "Task created successfully"
          },
          "400": {
            "description": "Invalid request"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ],
        "summary": "Create a new task",
        "tags": [
          "tasks"
        ]
      }
    },
    "/tasks/{id}": {
      "delete": {
        "operationId": "delete_task",
        "parameters": [
          {
            "description": "Task ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Task deleted successfully"
          },
          "400": {
            "description": "Invalid request"
          },
          "404": {
            "description": "Task not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ],
        "summary": "Delete a task",
        "tags": [
          "tasks"
        ]
      },
      "get": {
        "operationId": "get_task",
        "parameters": [
          {
            "description": "Task ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            },
            "description": "Task found"
          },
          "400": {
            "description": "Invalid task ID"
          },
          "404": {
            "description": "Task not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ],
        "summary": "Get a task by ID",
        "tags": [
          "tasks"
        ]
      },
      "put": {
        "operationId": "update_task",
        "parameters": [
          {
            "description": "Task ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTask"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            },
            "description": "Task updated successfully"
          },
          "400": {
            "description": "Invalid request"
          },
          "404": {
            "description": "Task not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ],
        "summary": "Update a task",
        "tags": [
          "tasks"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "API base path",
      "url": "/api"
    }
  ],
...
}