# libs/contracts/tasks/types/TaskPriority.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * Task priority levels
 */
export type TaskPriority = "low" | "medium" | "high" | "urgent";