# libs/contracts/tasks/types/UpdateTask.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TaskPriority } from "./TaskPriority";
import type { TaskStatus } from "./TaskStatus";

/**
 * DTO for updating an existing task
 */
export type UpdateTask = { title: string | null, description: string | null, completed: boolean | null, project_id: string | null | null, priority: TaskPriority | null, status: TaskStatus | null, due_date: string | null | null, };